requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
gunicorn==20.0.4
pymongo==3.11
python-dotenv>=0.19.2
authlib>=1.0
# auth0-python==4.4.0
